test: address partition reader review nits
