+++ lance-spark-base_2.12/src/test/java/org/lance/spark/read/LanceColumnarPartitionReaderTest.java
@@ -14,19 +14,27 @@
 package org.lance.spark.read;
 
 import org.lance.ipc.ColumnOrdering;
+import org.lance.namespace.LanceNamespace;
+import org.lance.namespace.model.DescribeTableRequest;
+import org.lance.namespace.model.DescribeTableResponse;
+import org.lance.spark.LanceSparkReadOptions;
 import org.lance.spark.TestUtils;
 import org.lance.spark.utils.Optional;
 
+import org.apache.arrow.memory.BufferAllocator;
 import org.apache.spark.sql.vectorized.ColumnarBatch;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class LanceColumnarPartitionReaderTest {
@@ -231,6 +239,157 @@ public class LanceColumnarPartitionReaderTest {
       assertDoesNotThrow(reader::close, "second close() must be a no-op");
     } finally {
       reader.close();
+    }
+  }
+
+  @Test
+  public void testPartitionReaderSkipsNamespaceWhenExecutorCredentialRefreshDisabled()
+      throws Exception {
+    RecordingNamespace.reset();
+    LanceInputPartition partition =
+        namespacePartition(
+            TestUtils.TestTable1Config.datasetUri,
+            Collections.singletonList(0),
+            "testPartitionReaderSkipsNamespaceWhenExecutorCredentialRefreshDisabled",
+            false);
+    try (LanceColumnarPartitionReader reader = new LanceColumnarPartitionReader(partition)) {
+      while (reader.next()) {
+        reader.get().close();
+      }
+    }
+
+    assertNull(
+        partition.getReadOptions().getNamespace(),
+        "executor_credential_refresh=false must skip namespace rebuild on the executor");
+    assertEquals(
+        0,
+        RecordingNamespace.INITIALIZE_CALLS.get(),
+        "executor_credential_refresh=false must not load or initialize the namespace impl");
+  }
+
+  @Test
+  public void testPartitionReaderReusesAndClosesExecutorNamespace() throws Exception {
+    RecordingNamespace.reset();
+    LanceInputPartition partition =
+        namespacePartition(
+            TestUtils.TestTable1Config.datasetUri,
+            Arrays.asList(0, 1),
+            "testPartitionReaderReusesAndClosesExecutorNamespace",
+            true);
+
+    LanceColumnarPartitionReader reader = new LanceColumnarPartitionReader(partition);
+    int rowsRead = 0;
+    try {
+      while (reader.next()) {
+        ColumnarBatch batch = reader.get();
+        assertNotNull(batch);
+        rowsRead += batch.numRows();
+        batch.close();
+      }
+    } finally {
+      reader.close();
+    }
+    reader.close();
+
+    assertEquals(TestUtils.TestTable1Config.expectedValues.size(), rowsRead);
+    assertEquals(
+        1,
+        RecordingNamespace.INITIALIZE_CALLS.get(),
+        "all fragments in one Spark task must reuse one namespace client");
+    assertEquals(
+        1,
+        RecordingNamespace.CLOSE_CALLS.get(),
+        "the task-owned namespace client must be closed exactly once");
+  }
+
+  @Test
+  public void testPartitionReaderClosesExecutorNamespaceWhenFragmentOpenFails() throws Exception {
+    RecordingNamespace.reset();
+    LanceInputPartition partition =
+        namespacePartition(
+            TestUtils.TestTable1Config.datasetUri,
+            Collections.singletonList(999999),
+            "testPartitionReaderClosesExecutorNamespaceWhenFragmentOpenFails",
+            true);
+    LanceColumnarPartitionReader reader = new LanceColumnarPartitionReader(partition);
+
+    RuntimeException failure = assertThrows(RuntimeException.class, reader::next);
+
+    assertNotNull(failure.getCause());
+    assertEquals(1, RecordingNamespace.INITIALIZE_CALLS.get());
+    assertEquals(
+        1,
+        RecordingNamespace.CLOSE_CALLS.get(),
+        "namespace initialization must be rolled back when fragment open fails");
+    reader.close();
+    assertEquals(
+        1,
+        RecordingNamespace.CLOSE_CALLS.get(),
+        "close after failed initialization cleanup must be idempotent");
+  }
+
+  private LanceInputPartition namespacePartition(
+      String location, List<Integer> fragments, String scanId, boolean executorCredentialRefresh) {
+    LanceSparkReadOptions readOptions =
+        LanceSparkReadOptions.builder()
+            .datasetUri(location)
+            .tableId(Collections.singletonList(TestUtils.TestTable1Config.datasetName))
+            .executorCredentialRefresh(executorCredentialRefresh)
+            .build();
+    return new LanceInputPartition(
+        TestUtils.TestTable1Config.schema,
+        0,
+        new LanceSplit(fragments),
+        readOptions,
+        Optional.empty(),
+        Optional.empty(),
+        Optional.empty(),
+        Optional.empty(),
+        Optional.empty(),
+        scanId,
+        Collections.emptyMap(),
+        RecordingNamespace.class.getName(),
+        Collections.singletonMap("location", location),
+        null);
+  }
+
+  /**
+   * Public, top-level-by-FQCN, no-arg {@link LanceNamespace} so that {@link
+   * LanceNamespace#connect(String, Map, BufferAllocator)} can resolve it via {@code Class.forName}
+   * during executor-side namespace initialization.
+   */
+  public static class RecordingNamespace implements LanceNamespace, AutoCloseable {
+    static final AtomicInteger INITIALIZE_CALLS = new AtomicInteger();
+    static final AtomicInteger CLOSE_CALLS = new AtomicInteger();
+
+    private String location;
+
+    public RecordingNamespace() {}
+
+    static void reset() {
+      INITIALIZE_CALLS.set(0);
+      CLOSE_CALLS.set(0);
+    }
+
+    @Override
+    public void initialize(Map<String, String> properties, BufferAllocator allocator) {
+      INITIALIZE_CALLS.incrementAndGet();
+      location = properties.get("location");
+    }
+
+    @Override
+    public String namespaceId() {
+      return "recording";
+    }
+
+    @Override
+    public DescribeTableResponse describeTable(DescribeTableRequest request) {
+      return new DescribeTableResponse().location(location);
+    }
+
+    @Override
+    public void close() {
+      CLOSE_CALLS.incrementAndGet();
     }
   }
 }
